Move tracing classes into wvlet.airframe.tracing
